--- Assets/Plugins/iOS.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: f3f3ab55ec6f8fe4a8ad1581c66be646
-folderAsset: yes
-DefaultImporter:
-  userData:
