+++ Assets/Scripts/Tool/DataUtilityManager.cs
@@ -6,11 +6,10 @@ using UnityEngine.Networking;
 
 public class DataUtilityManager
 {
-#if UNITY_EDITOR || UNITY_STANDALONE_WIN
-    public static string m_platform = "Windows";//当前平台
+#if UNITY_EDITOR
+    public static string m_platform = "Windows";
     public static string m_localRootPath = Application.streamingAssetsPath.Replace("Assets/StreamingAssets", "");//本地数据根目录
-
-#elif UNITY_ANDROID
+#else
     public static string m_platform = "Android";
     public static string m_localRootPath = Application.persistentDataPath + "/";
 #endif
