--- CatalogueFiles/Android/CatalogueFile.txt
@@ -1,1 +1,0 @@
-{"Bin/Config/Client/Player.bin":"0E98D7CA9E9BC41024EFBD68403F4E15","Bin/Config/Client/RoleRune.bin":"09783F81C213D5D5FDBAD22E1990E658","Bin/Atlas/Atlas00.bin":"601C44DBC16B47188D9C88889679E96C","Bin/Atlas/Atlas01.bin":"71ABD29623FD29554124D6347D56AD17","Bin/Atlas/Atlas02.bin":"3C88D5A68C6D9DD8B57BF178D24D042F","Lua/ToLua/Build.bat":"796C40AA00EAB69C48224476A2626994","Lua/ToLua/event.lua":"6D68598F5EEDC42B49361933D2E8FAD7","Lua/ToLua/list.lua":"6F1F9DBC54A03A284FA9027E23EEEE9F","Lua/ToLua/ltn12.lua":"D6101EB65C945775E74A379A5999B965","Lua/ToLua/mime.lua":"4C0FFE434F8AD9B8403FF3AD260002E1","Lua/ToLua/slot.lua":"BE8AA7A68CCAFD41F4C3A224F723DC1A","Lua/ToLua/socket.lua":"39460CDE4B85FBD37308930A49FED8A3","Lua/ToLua/tolua.lua":"702D8B57ACD700CD3323C6A7B88C254B","Lua/ToLua/typeof.lua":"680BA0A08B29509E5CD70E9976490901","Lua/ToLua/cjson/util.lua":"FDC0F01F3DB1E6FB373005EBA3928BAC","Lua/ToLua/jit/bc.lua":"071540D8AF4572EAAFA1457089A05BDA","Lua/ToLua/jit/bcsave.lua":"334068A533FFC821E08A281F0D411C7F","Lua/ToLua/jit/dis_arm.lua":"4B53131209C47C8FD27ABD6E7386E399","Lua/ToLua/jit/dis_arm64.lua":"2D0A9341B49774D3A721B51F409AC1B4","Lua/ToLua/jit/dis_arm64be.lua":"FD0E1319BB5D2F2534C6C679D68831BD","Lua/ToLua/jit/dis_mips.lua":"4ADD05AF1AB1614D2F327B5ADABDBBF0","Lua/ToLua/jit/dis_mips64.lua":"530D02AEAE18309E31F40EBC349EE532","Lua/ToLua/jit/dis_mips64el.lua":"FA0874005866BC035B7FF773F41AD91E","Lua/ToLua/jit/dis_mipsel.lua":"9002A02F5AAB7E5D1A15DECD954329F9","Lua/ToLua/jit/dis_ppc.lua":"DC53BC2F8AED87185629E3498BAD5909","Lua/ToLua/jit/dis_x64.lua":"9BDCF971745800E11A5BCD5E52876D3D","Lua/ToLua/jit/dis_x86.lua":"333F0ACEBA7587909DE2AA4F4E40EB73","Lua/ToLua/jit/dump.lua":"857E22765183E05DB50928827AE36956","Lua/ToLua/jit/p.lua":"65721CD3491E42F895D10A5F9859AA34","Lua/ToLua/jit/v.lua":"917F6C3A72446AF8EBDD150E9C32E850","Lua/ToLua/jit/zone.lua":"A0028573D5D34849A397030C8F50147C","Lua/ToLua/lpeg/re.lua":"AEDDDBE68E6B12B4A61D77CCADA15594","Lua/ToLua/misc/functions.lua":"B5A98BCF078842E475FA53DB86762524","Lua/ToLua/misc/strict.lua":"41A874EF9D517F384CBE80732A55D61E","Lua/ToLua/misc/utf8.lua":"8D9CC2AA4E6EBC7830228DBCB163C230","Lua/ToLua/protobuf/containers.lua":"5D5E79D72ACD578E05D4874D3C30CDBB","Lua/ToLua/protobuf/decoder.lua":"15871DCA255CAA53A82F839882122606","Lua/ToLua/protobuf/descriptor.lua":"20304A3E1A97075388F5F8579A027EEE","Lua/ToLua/protobuf/encoder.lua":"15A6DFC964AA67A8C053037AF6D91BAB","Lua/ToLua/protobuf/listener.lua":"258ED7863AEC5CC20B3C2C62832BB81D","Lua/ToLua/protobuf/protobuf.lua":"4BAC4F19113DB99E7A4722CC44C1A5DF","Lua/ToLua/protobuf/text_format.lua":"356DDAB80D5BC75ED4F581A5DF6F92F0","Lua/ToLua/protobuf/type_checkers.lua":"99A59C0F955D6756D42964D101F8611B","Lua/ToLua/protobuf/wire_format.lua":"C2D9AF63A6C4100B0749256E617F31C0","Lua/ToLua/socket/ftp.lua":"BC8130A9FBA0DBD1E828F286F5C21A22","Lua/ToLua/socket/headers.lua":"7D2FDDFB288B3424C706FDFE15F6BED5","Lua/ToLua/socket/http.lua":"BB0706774F7CC15ED96A06282446C761","Lua/ToLua/socket/mbox.lua":"E0BE88BF381745AE754BD4E479C1CEA2","Lua/ToLua/socket/smtp.lua":"8BB62043596F1232D0083AEF3F189A7F","Lua/ToLua/socket/tp.lua":"4AF09F1A5B0447192F8EEACF8FB5CA85","Lua/ToLua/socket/url.lua":"D513712FBE693FD3AD4DB9A1E2019FF0","Lua/ToLua/System/coroutine.lua":"531FFA03E5B8C3F5D739F2A486A4BDEB","Lua/ToLua/System/Timer.lua":"8EA8B70EE4F93908BDACFCC99CC73C78","Lua/ToLua/System/ValueType.lua":"30F0B769E327FF7BC320A1513E076C10","Lua/ToLua/System/Injection/InjectionBridgeInfo.lua":"3CDEB8C5F4C7F8CEB0FA5EAE6E3A1EA6","Lua/ToLua/System/Injection/LuaInjectionBus.lua":"BF2CE47024E7B5BA40C36C289BEA7729","Lua/ToLua/System/Injection/LuaInjectionStation.lua":"AE0E21FCE0AE0099FA323CF66A240BDB","Lua/ToLua/System/Reflection/BindingFlags.lua":"A23E75A97E18085D2D06A3D27643CFF0","Lua/ToLua/UnityEngine/Bounds.lua":"6379B9A1F85BAAFDE584FE99049159FB","Lua/ToLua/UnityEngine/Color.lua":"7D5F0992A7CBF3FD3A7FD54F71D48C59","Lua/ToLua/UnityEngine/LayerMask.lua":"BD3A93C2564CC3B9AE252C1B36A1A3C2","Lua/ToLua/UnityEngine/Mathf.lua":"420644580DE2FDA25B01D460B9ECA399","Lua/ToLua/UnityEngine/Plane.lua":"2247B8080B798534018390A3EC2E9EB1","Lua/ToLua/UnityEngine/Profiler.lua":"5AF614F04E585CF80D245291DEA60B0C","Lua/ToLua/UnityEngine/Quaternion.lua":"D87B3A33B551B8237D61BAD16E9A0451","Lua/ToLua/UnityEngine/Ray.lua":"7FAA634FDBFF0B390CC1949501F0979F","Lua/ToLua/UnityEngine/RaycastHit.lua":"8697148EEC19DC6E9EE71E50C6C49658","Lua/ToLua/UnityEngine/Time.lua":"B27C2618480BA321E3DC156E77E2DE4A","Lua/ToLua/UnityEngine/Touch.lua":"9FA7ADE85F1384CDC78BA81CC853DE6F","Lua/ToLua/UnityEngine/Vector2.lua":"D4C0F12336B81A67A6D89F89063446ED","Lua/ToLua/UnityEngine/Vector3.lua":"877140760BACD5CBB973866E9F19694E","Lua/ToLua/UnityEngine/Vector4.lua":"DBAA8D6C335B6BA48F4EC2C4A376BAFA","Lua/UI/LoginPanel.lua":"8CD1D85E7A0E2A1CFE95868791FDE012","Lua/Workflow/ConfigDataManager.lua":"0A8B0D05168CD553D5753354290BF7E6","Lua/Workflow/Launcher.lua":"64AC509F7727E40045736F84505F0FF1","Lua/Workflow/Main.lua":"270EC1ACE1BB7B77A85FFBDE5CA16509","Lua/Workflow/PrefabClass.lua":"8787E2B3BFB6D2B5E8CC1640837CBE6B","Lua/Workflow/ReflectUtil.lua":"171D9AFBD08B3DF296BF89307D036F61","AssetBundles/Android/Android.mainbundle":"C716A4623F5D5F0040A0E4F1F699EF88","AssetBundles/Android/Android.manifest":"5502C95C5EE2574B5F732D85D380B961","AssetBundles/Android/animtion/play.anim_ab":"929F912D9A962A8E0103FCA9A8BDC111","AssetBundles/Android/animtion/play.anim_ab.manifest":"1F87FC760C0D5C833BED1CC427EBF37E","AssetBundles/Android/atlas/atlas00.atlas_ab":"7F59F0FE500BBFF6EA44AB75F188C928","AssetBundles/Android/atlas/atlas00.atlas_ab.manifest":"F1506DB2DFDD795FD9CF9E22DE0B98BA","AssetBundles/Android/atlas/atlas01.atlas_ab":"4374B91545D3D17BF33614C75BFA5C42","AssetBundles/Android/atlas/atlas01.atlas_ab.manifest":"5B34A5D6392873453E19A0F6D5520624","AssetBundles/Android/atlas/atlas02.atlas_ab":"477F62F66521E508DE4F13AF58273F1C","AssetBundles/Android/atlas/atlas02.atlas_ab.manifest":"AFC826BEA0BF19A49E7244DD920A7CE7","AssetBundles/Android/atlas/bg_hghg_wdtx1.atlas_ab":"CB195FCAA06F3E1E3547D8BDD221BB70","AssetBundles/Android/atlas/bg_hghg_wdtx1.atlas_ab.manifest":"8037DE63BB241B5DEEA2682AE2CAD1D6","AssetBundles/Android/atlas/bg_hghg_wdtx2.atlas_ab":"F9349F8159A0CE4D98098C4B48A63199","AssetBundles/Android/atlas/bg_hghg_wdtx2.atlas_ab.manifest":"A77CE82DEE27EAA00D3F5B6DFDEF639C","AssetBundles/Android/atlas/bg_hghg_wdtx3.atlas_ab":"FB7C3B8D0AF4E26CC11086475EC96594","AssetBundles/Android/atlas/bg_hghg_wdtx3.atlas_ab.manifest":"282C6A13879A73F9116FA83A032DAE9E","AssetBundles/Android/audios/music.audio_ab":"745C7FEDFA6FEF8693E50572340E6EBC","AssetBundles/Android/audios/music.audio_ab.manifest":"5CE008CE2DFBB7B2E5F0997CEF97EFB1","AssetBundles/Android/materials/grayscale/grayscalematerial.mat_ab":"511408D855741FE24513986697CDEE8C","AssetBundles/Android/materials/grayscale/grayscalematerial.mat_ab.manifest":"009CDDFF3B475EEFC395B8CBD8D8FDC9","AssetBundles/Android/materials/uimaskchildren/uimaskchildrenmaterial.mat_ab":"95895E79014F97F6D30EDABD3A647C7F","AssetBundles/Android/materials/uimaskchildren/uimaskchildrenmaterial.mat_ab.manifest":"58DE4B424EC759090C54AEA9DBDB0DA9","AssetBundles/Android/materials/uimaskgrayscale/uimaskgrayscalematerial.mat_ab":"FE3FAA939A33C475BCE6A44E2CFDAA18","AssetBundles/Android/materials/uimaskgrayscale/uimaskgrayscalematerial.mat_ab.manifest":"003DE9ADCFC76975496FEA51A117FD6B","AssetBundles/Android/materials/uimaskparent/uimaskparentmaterial.mat_ab":"6B657860CC5FD982A0DB3D5EF015E649","AssetBundles/Android/materials/uimaskparent/uimaskparentmaterial.mat_ab.manifest":"DC932EC7D0B601467E15D4E9AF12C605","AssetBundles/Android/prefabs/ui/workflow/loginpanel.prefab_ab":"ACD4B9CA9311FA181E039895196B8352","AssetBundles/Android/prefabs/ui/workflow/loginpanel.prefab_ab.manifest":"6F8DAE3B5038F9ADD428319053633AF1"}
